backend/requirements.txt
asgiref==3.3.1
certifi==2020.12.5
Django==3.1.6
django-cors-headers==3.10.0
django-crispy-forms==1.11.0
django-mathfilters==1.0.0
djangorestframework==3.12.4
mysqlclient==2.0.3
pytz==2021.1
sqlparse==0.4.1
